fix: float32 match_idx overflow

+++ jina/drivers/score.py
@@ -23,7 +23,8 @@ class Chunk2DocScoreDriver(BaseExecutableDriver):
                     query_chunk_meta[c.chunk_id] = pb_obj2dict(c, exec.required_keys)
                     match_chunk_meta[k.match_chunk.chunk_id] = pb_obj2dict(k.match_chunk, exec.required_keys)
 
-            match_idx = np.array(match_idx, dtype=np.float32)
+            # np.uint32 uses 32 bits. np.float32 uses 23 bit mantissa, so integer greater than 2^23 will have their least significant bits truncated. For example, 4294967294 will be truncated to 4294967300. np.float64 uses 52 bits mantissa, which avoids the problem.
+            match_idx = np.array(match_idx, dtype=np.float64)
 
             doc_idx = self.exec_fn(match_idx, query_chunk_meta, match_chunk_meta)
